services/simulator/pyproject.toml
[project]
name = "disessa-simulator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mujoco>=3.1.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
]
